% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregateByStationIDAndYear.R
\name{AggregateByStationIDAndYear}
\alias{AggregateByStationIDAndYear}
\title{Aggregate the complete time.series by STATIONS_ID and Day/Month}
\usage{
AggregateByStationIDAndYear(DFToAggregate, AggregationFunction = "mean")
}
\arguments{
\item{AggregationFunction}{character. Aggregation function. Should be the arithmetic mean, median etc..
default: `mean`}
}
\value{
data.frame. Aggregated `TMK` and `RSK` values by STATIONS_ID and Day/Month
}
\description{
Aggregate the complete time.series by STATIONS_ID and Day/Month
}
