% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDistanceBetwLongAndLat.R
\name{CalculateDistanceBetwLongAndLat}
\alias{CalculateDistanceBetwLongAndLat}
\title{Distance between lat-long coordinates}
\usage{
CalculateDistanceBetwLongAndLat(
  ObjectDF,
  ObjectLatName,
  ObjectLongName,
  StationDF,
  StationLatName,
  StationLongName,
  RadiusAroundObject = 100,
  RadiusOfEarth = 6371
)
}
\arguments{
\item{RadiusOfEarth}{radius of the earth. Could be given in miles. DEFAULT: 6371 (km)}

\item{ObjectLat}{numeric. Latitude (North/South) coordinates in decimal degrees}

\item{ObjectLong}{numeric. Longitude (East/West) coordinates in decimal degrees}

\item{DFLongLat}{data.frame. data.frame with the columns `LAT` and `LONG`}

\item{i}{Integer: Index element against which all coordinate pairs
are computed. DEFAULT: 1}
}
\value{
Vector with distance(s) in km (or units of `r`, if `r` is changed)
}
\description{
Distance between lat-long coordinates
}
\details{
: Great-circle distance between points at lat-long coordinates.
The core is a copy of OSMscale::earthDist Version 0.5.3 (2017-04-19).
<https://github.com/brry/OSMscale/blob/master/R/earthDist.R#L57-L102>.
No checks of coordinates etc.
}
\examples{
PropertyData <- data.frame(Proj_key = c(1021, 1378, 1672, 1732, 1862),
                           ObjectName = c("Helogland (Reede)", "Wyk auf Föhr (Alte Mole)",
                                          "Konstanz (Hafen)", "Greetsiel (Hafen)",
                                          "Dornumersiel Hafen"),
                           LATITUDE = c(54.179837, 54.692713, 47.659868, 53.50214, 53.68001),
                           LONGITUDE = c(7.892211, 8.573618, 9.178899, 7.09950, 7.48596))

StationData <- data.frame(StationName = c("WARNEMUENDE", "JEVER", "LEUCHTTURM KIEL", "NAObject", "NORDERNEY"),
                          LAT =  c(54.11, 53.32, 54.30, NA, 53.43),
                          LONG = c(12.05,  7.53, 10.16, NA,  7.09))

CalculateDistanceBetwLongAndLat(ObjectDF = PropertyData,
                                ObjectLatName = "LATITUDE",
                                ObjectLongName = "LONGITUDE",
                                StationDF = StationData,
                                StationLatName = "LAT",
                                StationLongName = "LONG",
                                RadiusAroundObject = 100,
                                RadiusOfEarth = 6371)
}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2016 + Jan 2017.
        Thank you very much!
        Angle formula from Diercke Weltatlas 1996, Page 245
}
