% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckIfForecastingDataIsAvailable.R
\name{CheckIfForecastingDataIsAvailable}
\alias{CheckIfForecastingDataIsAvailable}
\title{Check if forecasting data is available}
\usage{
CheckIfForecastingDataIsAvailable(
  DWDForecastingDatList = DWDForecastingDatList,
  DWDForecastingMetaDataFromKMLObject = DWDForecastingMetaDataFromKMLObject,
  DistancesObjectWXStations = DistancesObjectWXStations
)
}
\arguments{
\item{DistancesObjectWXStations}{}
}
\value{
data.frame. Weatherstation metadata.
}
\description{
Check if forecasting data is available
}
