% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckIfWeatherDataIsComplete.R
\name{CheckIfWeatherDataIsComplete}
\alias{CheckIfWeatherDataIsComplete}
\title{Checks if a DWD historical weather data.frame has complete data in between a given start and end date}
\usage{
CheckIfWeatherDataIsComplete(
  HistoricalWeatherDataFrameToTest,
  StartYear = 2022,
  EndYear = 2022,
  silent = TRUE
)
}
\arguments{
\item{HistoricalWeatherDataFrameToTest}{data.frame.}

\item{StartYear}{integer. default: `2022`}

\item{EndYear}{integer. default: `2022`}

\item{silent}{boolean. Should messages be print? default: `TRUE`}
}
\value{
list.
}
\description{
Checks if a DWD historical weather data.frame has complete data in between a given start and end date
}
\details{
Checks whole years and if data is available on every day; Start year starts at 1st of January, end of year
31st of December
}
\examples{
# Some weather station IDs
VectorIDsSomeWeatherStations <- c("02115", "01963", "02712")

# Get the data from the weather stations: historical and recent
HistoricalWeatherDataDF <- GetHistoricalDWDWeatherData(VectorWithIDsDWDWeatherStations = VectorIDsSomeWeatherStations)
# First six rows
head(HistoricalWeatherDataDF)
CheckIfWeatherDataIsComplete(HistoricalWeatherDataFrameToTest = HistoricalWeatherDataDF)
}
