% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEmptyTimeSeriesDF.R
\name{CreateEmptyTimeSeriesDF}
\alias{CreateEmptyTimeSeriesDF}
\title{Create an Empty Time Series data.frame between Start- and Endyear for each station}
\usage{
CreateEmptyTimeSeriesDF(StartYear = 2019, EndYear = 2022, StationIDs)
}
\arguments{
\item{StartYear}{integer. Should be the start year. Start year starts at 1st of January. default = `2019`}

\item{EndYear}{integer. Should be the end year. End year ends at 31st of December. default = `2022`}

\item{StationIDs}{character. Character vector with all DWD historical weather data station ids.}
}
\value{
data.frame. data.frame with the complete time series for each station id.
}
\description{
Create an Empty Time Series data.frame between Start- and Endyear for each station
}
\examples{
CreateEmptyTimeSeriesDF(StartYear = 2019, EndYear = 2022, StationIDs = c("05640", "05839"))
}
