% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDataFromForecastingKml.R
\name{ExtractDataFromForecastingKml}
\alias{ExtractDataFromForecastingKml}
\title{Extract data from kml forecasting file}
\usage{
ExtractDataFromForecastingKml(
  ExtractedDWDKMLFile,
  ParametersToExtract = c("RR1c", "ww", "T5cm", "TTT", "TN", "TX", "FF", "SunD", "N")
)
}
\arguments{
\item{DWDForecastXMLDatei}{XMLInternalDocument / XMLAbstractDocument. Extracted XML document from DWD kmz file.}
}
\value{
character. Variable names of the parameters to extract.
}
\description{
Extract data from kml forecasting file
}
\examples{
## Station K057
TestRequestDF <-
  data.frame(ID = "K057",
             URL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/K057/kml/MOSMIX_L_LATEST_K057.kmz")
# Request the DWD forecasting weather data
PathToFolderAndFile <- RequestMOSMIXLData(RequestURLDataFrame = TestRequestDF, ColumnWeatherID = "ID", ColumnForecastingURL = "URL")
# Extract XML
ExtractedForecastKMLObject <- UnpackDWDForecastFile(PathToFolderAndFile = PathToFolderAndFile)
class(ExtractedForecastKMLObject)
# Extract station data
ExtractedForecastData <- ExtractDWDForecastStationDataFromKML(ExtractedDWDKMLFile = ExtractedForecastKMLObject)
# Stationsdaten extrahieren
head(ExtractDataFromForecastingKml(ExtractedDWDKMLFile = ExtractedForecastKMLObject))
}
