% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDWDForecastStationDataFromKML.R
\name{ExtractDWDForecastStationDataFromKML}
\alias{ExtractDWDForecastStationDataFromKML}
\title{Extract DWD forecasting station data from extracted kml file}
\usage{
ExtractDWDForecastStationDataFromKML(ExtractedDWDKMLFile)
}
\arguments{
\item{DWDForecastXMLDatei}{XMLInternalDocument / XMLAbstractDocument. Extracted XML document from DWD kmz file.}
}
\value{
SpatialPointsDataFrame. Objekt with `coordinates`, `name`, `desc` and `alt`
}
\description{
Extract DWD forecasting station data from extracted kml file
}
\examples{
## Station K057
TestRequestDF <-
  data.frame(ID = "K057",
             URL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/K057/kml/MOSMIX_L_LATEST_K057.kmz")
# Request the DWD forecasting weather data
PathToFolderAndFile <- RequestMOSMIXLData(RequestURLDataFrame = TestRequestDF, ColumnWeatherID = "ID", ColumnForecastingURL = "URL")
# Extract XML
ExtractedForecastKMLObject <- UnpackDWDForecastFile(PathToFolderAndFile = PathToFolderAndFile)
class(ExtractedForecastKMLObject)
# Extract station data
ExtractDWDForecastStationDataFromKML(ExtractedDWDKMLFile = ExtractedForecastKMLObject)
}
