% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForecastDWDWeatherData.R
\name{ForecastDWDWeatherData}
\alias{ForecastDWDWeatherData}
\title{Title}
\usage{
ForecastDWDWeatherData(
  UseBuildInDWDForecastingMetaData = FALSE,
  ObjectDataDF,
  VarNameProjKeyInObjectDataDF = "Proj_key",
  VarNameObjectNameInObjectDataDF = "ObjectName",
  VarNameLATInObjectDataDF = "LATITUDE",
  VarNameLONGInObjectDataDF = "LONGITUDE",
  RadiusAroundObject = 40
)
}
\arguments{
\item{UseBuildInDWDForecastingMetaData}{}
}
\description{
Title
}
\examples{
PropertyData <- data.frame(Proj_key = c(1021, 1378, 1672, 1732, 1862),
                           ObjectName = c("Helogland (Reede)", "Wyk auf Föhr (Alte Mole)",
                                          "Konstanz (Hafen)", "Greetsiel (Hafen)",
                                          "Dornumersiel Hafen"),
                           LATITUDE = c(54.179837, 54.692713, 47.659868, 53.50214, 53.68001),
                           LONGITUDE = c(7.892211, 8.573618, 9.178899, 7.09950, 7.48596))

ForecastDWDWeatherData(UseBuildInDWDForecastingMetaData = FALSE,
                       ObjectDataDF = PropertyData,
                       VarNameProjKeyInObjectDataDF = "Proj_key",
                       VarNameObjectNameInObjectDataDF = "ObjectName",
                       VarNameLATInObjectDataDF = "LATITUDE",
                       VarNameLONGInObjectDataDF = "LONGITUDE")
}
