% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateCompleteTimeSeriesDataFrame.R
\name{GenerateCompleteTimeSeriesDataFrame}
\alias{GenerateCompleteTimeSeriesDataFrame}
\title{Generates a complete time series data frame between start and end year}
\usage{
GenerateCompleteTimeSeriesDataFrame(
  StartYear = StartYear,
  EndYear = EndYear,
  HistDataList = ListWithResults$HistoricalWeatherDataList
)
}
\arguments{
\item{StartYear}{integer. Start year.}

\item{EndYear}{integer. End year.}

\item{HistDataList}{list. List with data.frames for every DWD historical weather station.}
}
\value{
list. List with DWD historical weather station with a complete set of dates between startdate and enddate.
}
\description{
Generates a complete time series data frame between start and end year
}
\details{
Checks if the data.frame with historical weather data from DWD is complete due to all dates between
start and end date. If in the data.frame dates are missing, they will be added. Corresponding values of `RSK` and
`TMK` will be set to NA for these missing dates. NAs can be imputet in an extra step.
}
\examples{
GenerateCompleteTimeSeriesDataFrame(StartYear = StartYear, EndYear = EndYear,
                                    HistDataList = ListWithResults$HistoricalWeatherDataList)
}
