% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateForcastingURLs.R
\name{GenerateForcastingURLs}
\alias{GenerateForcastingURLs}
\title{Creates a URL to query the DWD Forecasting weather data based on the ID of the weather station}
\usage{
GenerateForcastingURLs(IDWeatherStation)
}
\arguments{
\item{IDWeatherStation}{character. Weather station ID as character.}
}
\value{
data.frame. ID of the weather station with URL.
}
\description{
Creates a URL to query the DWD Forecasting weather data based on the ID of the weather station
}
\examples{
GenerateForcastingURLs(IDWeatherStation = "E273")
GenerateForcastingURLs(IDWeatherStation = "E355")
}
