% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDWDForecastingMetaData.R
\name{GetDWDForecastingMetaData}
\alias{GetDWDForecastingMetaData}
\title{Obtaining the metadata of the DWD Forecasting (MOSMIXL) weather stations}
\usage{
GetDWDForecastingMetaData()
}
\value{
data.frame. DWD Forecasting weather stations metadata.
}
\description{
The raw data are obtained from an opendata server of the DWD. The obtained data are in fixed format.
}
\examples{
DWDForecastingMetaData <- GetDWDForecastingMetaData()
}
