% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDWDMetaData.R
\name{GetDWDMetaData}
\alias{GetDWDMetaData}
\title{Obtaining the metadata of the DWD weather stations}
\usage{
GetDWDMetaData(OnlyCurrentDate = FALSE)
}
\arguments{
\item{OnlyCurrentDate}{bool. Should only weather stations be displayed that provide weather data until
today? default: `FALSE`}
}
\value{
data.frame. DWD weather stations metadata.
}
\description{
Obtaining the metadata of the DWD weather stations
}
\examples{

# Obtain all weather stations
DWDHistoricalMetaDataAllStations <- GetDWDMetaData(OnlyCurrentDate = FALSE)
head(DWDHistoricalMetaDataAllStations)
str(DWDHistoricalMetaDataAllStations)

# Obtain only weather stations that provide data until yesterday
DWDHistoricalMetaData <- GetDWDMetaData(OnlyCurrentDate = TRUE)
head(DWDHistoricalMetaData)
str(DWDHistoricalMetaData)
}
