% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHistoricalDWDWeatherData.R
\name{GetHistoricalDWDWeatherData}
\alias{GetHistoricalDWDWeatherData}
\title{Get historical DWD weather data based on nearby IDs}
\usage{
GetHistoricalDWDWeatherData(VectorWithIDsDWDWeatherStations)
}
\arguments{
\item{VectorWithIDsDWDWeatherStations}{integer. Vector with the IDs of the DWD weather stations}
}
\value{
data.frame. List with a match of the extracted data.frames of the historical data of each DWD
 weather station by ID from a list. The query returns a list of individual data.frames, which are then merged
 into a common data.frame.
}
\description{
Get historical DWD weather data based on nearby IDs
}
\examples{
# Some weather station IDs
VectorIDsSomeWeatherStations <- c("02115", "01963", "02712")

# Get the data from the weather stations: historical and recent
HistoricalWeatherDataDF <- GetHistoricalDWDWeatherData(VectorWithIDsDWDWeatherStations = VectorIDsSomeWeatherStations)
# First six rows
head(HistoricalWeatherDataDF)

# Checking the most recent entry
aggregate(x = HistoricalWeatherDataDF[,"MESS_DATUM", drop = F], by =  list(HistoricalWeatherDataDF$STATIONS_ID), max)

# Check the oldest entry
aggregate(x = HistoricalWeatherDataDF[,"MESS_DATUM", drop = F], by =  list(HistoricalWeatherDataDF$STATIONS_ID), min)
}
