% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistoricalDWDWeatherData.R
\name{HistoricalDWDWeatherData}
\alias{HistoricalDWDWeatherData}
\title{Main function: Get a data.frame with historical weather data from DWD}
\usage{
HistoricalDWDWeatherData(
  DataFrame = PropertyData.1,
  Proj_key = Proj_key,
  NAME = NAME,
  LATITUDE = LATITUDE,
  LONGITUDE = LONGITUDE,
  ExtractedRunningNo = 2,
  StartYear = 2018,
  EndYear = 2022,
  PrintMessages = TRUE,
  ThresholdNAs = 0.05
)
}
\arguments{
\item{ExtractedRunningNo}{integer. Extracted DWD weather station in downloaded list. It is the list with smallest
path to object coordinates. default: `2`}

\item{StartYear.}{integer. default: `2018`}

\item{EndYear.}{integer. default: `2022`}

\item{PrintMessages.}{boolean. default: `TRUE`}

\item{ThresholdNAs.}{double. default: `0.05`}
}
\value{
list. Returns a list with the different data.frames
}
\description{
Main function: Get a data.frame with historical weather data from DWD
}
\examples{
PropertyData.1 <- data.frame(Proj_key = c(1021, 1378, 1672),
                             NAME = c("Helogland (Reede)", "Wyk auf Föhr (Alte Mole)", "Konstanz (Hafen)"),
                             LATITUDE = c(54.179837, 54.692713, 47.659868),
                             LONGITUDE = c(7.892211, 8.573618, 9.178899))

HistDWDDF <- HistoricalDWDWeatherData(DataFrame = PropertyData.1,
                                      Proj_key = Proj_key,
                                      NAME = NAME,
                                      LATITUDE = LATITUDE,
                                      LONGITUDE = LONGITUDE,
                                      ExtractedRunningNo = 2,
                                      StartYear = 2018,
                                      EndYear = 2022,
                                      PrintMessages = TRUE,
                                      ThresholdNAs = 0.05)
}
