% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NearbyDWDStations.R
\name{NearbyDWDStations}
\alias{NearbyDWDStations}
\title{Identify nearby DWD stations}
\usage{
NearbyDWDStations(
  propertyData,
  latitude = "LATITUDE",
  longitude = "LONGITUDE",
  radius = 50,
  minDate = "2021-01-01"
)
}
\arguments{
\item{propertyData}{data.frame. Properties with longitude and latitude.}

\item{latitude}{character. Variable name latitude. default `LATITUDE`}

\item{longitude}{character. Variable name longitude. default `LONGITUDE`}

\item{radius}{integer. Radius in kilometers to be considered around longitude and latitude.}

\item{minDate}{character. Minimum date. Format: 'YY-MM-DD', e.g. '2021-12-25'. default: `2021-01-01`}
}
\value{
data.frame. Identified near-by stations with added information.
}
\description{
The first Nearby DWD station of a serviced object is always identified.
}
\examples{
# Example 1:
PropertyData.1 <- data.frame(Proj_key = c(1021, 1378, 1672),
                             NAME = c("Helogland (Reede)", "Wyk auf Föhr (Alte Mole)", "Konstanz (Hafen)"),
                             LATITUDE = c(54.179837, 54.692713, 47.659868),
                             LONGITUDE = c(7.892211, 8.573618, 9.178899))

NearbyDWDStationsDataFrame <- NearbyDWDStations(propertyData = PropertyData.1, radius = 50)
}
