% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefreshStationIDs.R
\name{RefreshStationIDs}
\alias{RefreshStationIDs}
\title{Update DWD historical weather station IDs if NAs above threshold}
\usage{
RefreshStationIDs(
  ValidationResultsDF,
  NearbyDWDStationsDataFrame,
  ExtractedRunningNo,
  IDExtractedWeatherStations
)
}
\arguments{
\item{IDExtractedWeatherStations}{}
}
\value{
character. Character vector with update of station ids to get historical DWD weather data for
}
\description{
Update DWD historical weather station IDs if NAs above threshold
}
