% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RequestMOSMIXLData.R
\name{RequestMOSMIXLData}
\alias{RequestMOSMIXLData}
\title{Download the forecast weather data of the DWD}
\usage{
RequestMOSMIXLData(
  RequestURLDataFrame,
  ColumnWeatherID = "ID",
  ColumnForecastingURL = "ForecastingURL"
)
}
\arguments{
\item{RequestURLDataFrame}{data.frame. data.frame with the query URL and the DWD weather station ID}

\item{ColumnWeatherID}{character. DWD forecasting weatherstation ID}

\item{ColumnForecastingURL}{character. URL request.}
}
\value{
character. Returns the path and folder of the download.
}
\description{
Download the forecast weather data of the DWD
}
\examples{
TestRequestDF = data.frame(
  ID = "11120",
  ForecastingURL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/11120/kml/MOSMIX_L_LATEST_11120.kmz")
TestRequestDF
# Request DWD forecast weather data
RequestMOSMIXLData(RequestURLDataFrame = TestRequestDF)
}
