% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnpackDWDForecastFile.R
\name{UnpackDWDForecastFile}
\alias{UnpackDWDForecastFile}
\title{Unpack DWD Forecasting File - kmz-Format - and return as XML document}
\usage{
UnpackDWDForecastFile(PathToFolderAndFile)
}
\arguments{
\item{PathToFolderAndFile}{character. Folder where the kmz file is stored.}
}
\value{
XMLInternalDocument / XMLAbstractDocument. Generated XML document for further processing.
}
\description{
Unpack DWD Forecasting File - kmz-Format - and return as XML document
}
\examples{
## Station K057
TestRequestDF <-
  data.frame(ID = "K057",
             URL = "https://opendata.dwd.de/weather/local_forecasts/mos/MOSMIX_L/single_stations/K057/kml/MOSMIX_L_LATEST_K057.kmz")
# Request the DWD forecasting weather data
PathToFolderAndFile <- RequestMOSMIXLData(RequestURLDataFrame = TestRequestDF, ColumnWeatherID = "ID", ColumnForecastingURL = "URL")
# Extract XML
ExtrahierteForecastXML <- UnpackDWDForecastFile(PathToFolderAndFile = PathToFolderAndFile)
class(ExtrahierteForecastXML)
}
