% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImputeTimeSeries.R
\name{ImputeTimeSeries}
\alias{ImputeTimeSeries}
\title{Impute time series with NA values to get complete data}
\usage{
ImputeTimeSeries(
  WXValidationDF = ListWithResults[["WXValidationDF"]]$ValidationAggrDFPercentageValues,
  TimeSeriesDF = ListWithResults[["HistoricalWeatherDataDFReduced"]]
)
}
\arguments{
\item{TimeSeriesDF}{}
}
\value{
data.frame. complete time.series data.frame.
}
\description{
Impute time series with NA values to get complete data
}
